requirements.txt
pandas==2.2.2
mysql-connector-python==8.4.0
python-dotenv==1.0.1